Move the tool functions for learning and disambiguation by the tonal form in differential_tone_coding.py and an adjustment for reordering lists.
