+++ disambiguation.py
@@ -17,230 +17,20 @@
 #
 
 
-import sys, re, codecs, glob, time, os
-import argparse
+import sys, re, codecs, glob, time, os, collections, argparse, itertools
 import formats,  grammar
-import collections
+from gdisamb import FileParser
 from ntgloss import Gloss
 from nltk.tag.crf import CRFTagger
-from gdisamb import FileParser
-from differential_tone_coding import is_a_good_code, code_seperator, get_features_customised, chunking, encoder_tones, repr, get_features_customised_tone, code_dispatcher, code_resort, mode_indicators
-import unicodedata
-import pycrfsuite
-import csv
 import nltk.tag.util
-import itertools
-from nltk.metrics.scores import accuracy
+import pycrfsuite
+from differential_tone_coding import get_duration, sampling, csv_export, unzip, encoder_tones, mode_indicators, marginal_tone, accuray2, get_sub_tone_code_of_sentence, accumulate_tone_code_of_dataset, reshape_tokens_as_sentnece, make_tokens_from_sentence, make_features_from_tokens
+import unicodedata
 import zipfile
 
 import codecs, sys
 sys.stdin = codecs.getreader('utf8')(sys.stdin)
 sys.stdout = codecs.getwriter('utf8')(sys.stdout)
-
-def marginal_tone(taggers, tnum, tokens, tag, token) :
-
-	enc = encoder_tones()
-	codes, syllabes = enc.differential_encode(token, tag.decode('utf-8'))
-
-	p = 0
-	snums = []
-	for i, token in enumerate(tokens) :
-		for j, syllabe in enumerate(token) :
-			if i == tnum :
-				snums.append(p)
-			p += 1
-
-	prob = 1
-	for phase, tagger in enumerate(taggers) :
-		for i in range(len(syllabes)) :
-			subcode = code_dispatcher(codes[i])[phase]
-			print subcode
-			prob *= tagger._tagger.marginal(subcode.encode('utf-8'), snums[i])
-
-	return prob
-
-def accuray2 (dataset1, dataset2, is_tone_mode = False) :
-	cnt_sucess = 0
-	cnt_fail = 0
-	if not is_tone_mode :
-		for sent1, sent2 in zip(dataset1, dataset2) :
-			for token1, token2 in zip(sent1, sent2) :
-				if token1 == token2 :
-					cnt_sucess += 1
-				else :
-					cnt_fail += 1
-
-	else :
-		for sent1, sent2 in zip(dataset1, dataset2) :
-			for token1, token2 in zip(sent1, sent2) :
-				is_identical = True
-				for syllabe1, syllabe2 in zip(token1, token2) :
-					if syllabe1 != syllabe2 : is_identical = False ; break;
-				if is_identical :
-					cnt_sucess += 1
-				else :
-					cnt_fail += 1
-
-	cnt_tot = cnt_sucess + cnt_fail
-	if not cnt_tot : return 0.0
-	else : return cnt_sucess / float(cnt_tot)
-
-def get_sub_tone_code_of_sentence (sentence, phase) :
-	labels = list()
-	for i, token in enumerate(sentence) :
-		label = list()
-		for j, syllabe_code in enumerate(token) :
-			syllabe, code = syllabe_code
-			subcode = code_dispatcher(code.decode('utf-8'))[phase].encode('utf-8')
-			label.append(subcode)
-		labels.append(label)
-	return labels
-
-def accumulate_tone_code_of_dataset (dataset_acc, dataset) :
-	for p, sent in enumerate(dataset_acc) :
-		for i, token in enumerate(sent) :
-			for j, syllabe_tag_acc in enumerate(token) :
-				syllabe_acc, tag_acc = syllabe_tag_acc
-				syllabe, tag = dataset[p][i][j]
-				if tag_acc and tag :
-					tag_acc += code_seperator.encode('utf-8') + tag
-				else :
-					tag_acc += tag
-				dataset_acc[p][i][j] = \
-					tuple([syllabe, code_resort(tag_acc.decode('utf-8')).encode('utf-8')])
-
-	return dataset_acc
-
-def reshape_tokens_as_sentnece(tokens, sentnece) :
-
-	ret = list()
-	n = 0
-	for i, token in enumerate(sentnece) :
-		tmp = list()
-		for j, syllabe in enumerate(token) :
-			tmp.append(tokens[n])
-			n += 1
-		ret.append(tmp)
-
-	return ret
-
-def make_tokens_from_sentence(sent, is_tone_mode = False) :
-	if is_tone_mode :
-		tokens = list()
-		labels = list()
-		for token in sent :
-			tokens.append(unzip(token)[0])
-			labels.append(unzip(token)[1])
-	else :
-		tokens = unzip(sent)[0]
-		labels = unzip(sent)[1]
-
-	return [tokens, labels]
-
-def make_features_from_tokens(tokens, phase = 0, is_tone_mode = False) :
-	if is_tone_mode :
-		features_syllabe = list()
-		for i, token in enumerate(tokens) :
-			feature = list()
-			for j, syllabe_code in enumerate(token) :
-				feature.append(get_features_customised_tone(tokens, i, j, phase))
-			features_syllabe.append(feature)
-		features = list(itertools.chain(*features_syllabe))
-	else :
-		features = list()
-		for i in range(len(tokens)) :
-			features.append(get_features_customised(tokens, i))
-	return features
-
-def inspector_tokens(gold_tokens, predicted_tokens) :
-	for x,y in zip(gold_tokens, predicted_tokens) :
-		try :
-			if x[1] != y[1] :
-				print x[0],":",x[1].decode('utf-8'),"->",y[1].decode('utf-8') # ,"(",len(x[1]), len(y[1]),")"
-			else :
-				print "*",x[0],":",x[1].decode('utf-8'),"->",y[1].decode('utf-8') # ,"(",len(x[1]), len(y[1]),")"
-		except :
-			print type(x[0]),":",type(x[1]),"->",type(y[1])
-
-def unzip(input) :
-	return [list(li) for li in zip(*input)]
-
-def csv_export(filename, gold_set, test_set, is_tone_mode = False):
-
-	if not is_tone_mode :
-		csvfile = codecs.open(filename, 'wb')
-		writer = csv.writer(csvfile)
-		writer.writerow(["Token", "Golden", "Predicted", "Same"])
-		for gold_sent, test_sent in zip(gold_set, test_set) :
-			for gold_token, test_token in zip(gold_sent, test_sent) :
-				token = gold_token[0]
-				gold_code = gold_token[1]
-				test_code = test_token[-1]
-				# print token, gold_code, test_code
-				sameCodes = (gold_code == test_code)
-
-				if not repr(token.encode('utf-8')) :
-					sameCodes = u''
-				row = [\
-				(token.encode('utf-8')), \
-				gold_code, \
-				test_code, \
-				sameCodes]
-				writer.writerow(row)
-		csvfile.close()
-	else :
-		csvfile = codecs.open(filename, 'wb')
-		writer = csv.writer(csvfile)
-		writer.writerow(["Token", \
-			"Golden Form","Predicted Form", \
-			"Golden code", "Predicted code", "Same"])
-		enc = encoder_tones()
-		for gold_sent, test_sent in zip(gold_set, test_set) :
-			for gold_token, test_token in zip(gold_sent, test_sent) :
-				gold_code = ''
-				test_code = ''
-				gold_form = ''
-				test_form = ''
-				token = ''
-				for gold_syllabe, test_syllabe in zip(gold_token, test_token) :
-					token += gold_syllabe[0]
-					gold_code += gold_syllabe[1]
-					test_code += test_syllabe[1]
-					gold_form += enc.differential_decode(gold_syllabe[0], gold_syllabe[1].decode('utf-8'))
-					test_form += enc.differential_decode(gold_syllabe[0], test_syllabe[1].decode('utf-8'))
-					sameCodes = (gold_code == test_code)
-					sameForms = (gold_form == test_form)
-				sameCodes = (gold_code == test_code)
-				sameForms = (gold_form == test_form)
-				if not repr(token.encode('utf-8')) :
-					sameCodes = u''
-				row = [\
-				(token.encode('utf-8')), \
-				repr(gold_form.encode('utf-8')), \
-				repr(test_form.encode('utf-8')), \
-				repr(gold_code, True), \
-				repr(test_code, True), \
-				sameCodes]
-				writer.writerow(row)
-		csvfile.close()
-
-def sampling(allsents, p, ratio = 1) :
-	train_set, eval_set = [], []
-	for i, sent in enumerate(allsents[0 : : int(1/float(ratio))]) :
-		p_approx = float(len(train_set) + 1) / float(len(eval_set) + len(train_set) + 1)
-		if p_approx <= p :
-			train_set.append(sent)
-		else:
-			eval_set.append(sent)
-	return [train_set, eval_set]
-
-def get_duration(t1_secs, t2_secs) :
-	secs = abs(t1_secs - t2_secs)
-	days = secs // 86400
-	hours = secs // 3600 - days * 24
-	minutes = secs // 60 - hours * 60 - days * 60 * 24
-	secondes = int(secs) % 60
-	return '{:>02.0f}:{:>02.0f}:{:>02.0f}:{:>02d}'.format(days, hours, minutes, secondes)
 
 def main():
 
@@ -482,7 +272,7 @@ def main():
 							except : tag = '' ; print option
 							prob = tagger._tagger.marginal(tag, tnum)
 							options.append((prob, option))
-						reordered_probs, reordered_options = unzip(sorted(options, reverse = True))
+						reordered_probs, reordered_options = unzip(sorted(options, key = lambda x : x[0], reverse = True))
 						if args.select :
 							prob_max = reordered_probs[0]
 							reordered_options = tuple([reordered_options[i] for i, p in enumerate(reordered_probs) if p >= prob_max])
@@ -524,13 +314,16 @@ def main():
 							except : tag = ''
 							prob = marginal_tone(taggers, tnum, tokens, tag, token.token)
 							options.append((prob, option))
-						reordered_probs, reordered_options = unzip(sorted(options, reverse = True))
+
+						reordered_probs, reordered_options = unzip(sorted(options, key = lambda x : x[0], reverse = True))
 						if args.select :
 							prob_max = reordered_probs[0]
 							reordered_options = tuple([reordered_options[i] for i, p in enumerate(reordered_probs) if p >= prob_max])
 						html_parser.glosses[snum][1][tnum] = reordered_options
 
-		try : html_parser.write(args.outfile)
+		try :
+			html_parser.write(args.outfile)
+			print "Disambiggated resulat for {} is saved in {}".format(args.infile,args.outfile)
 		except IOError: print "Error : unable to create the output file {} !".format(args.outfile)
 
 	else :
